fix(web): split files page components for react-doctor
One component per .tsx under pages/files; move pure helpers to files-shared.ts so Fast Refresh stays intact.
