+++ apps/web/src/pages/files/files-artifact-row-menu.tsx
@@ -1,10 +1,7 @@
 import type { ArtifactFile } from "@nakama/core/contract";
 import {
   Delete02Icon,
-  File02Icon,
   FileDownloadIcon,
-  Film02Icon,
-  Image02Icon,
   MoreHorizontalIcon,
 } from "hugeicons-react";
 import {
@@ -12,7 +9,6 @@ import {
   ArtifactSharePublishDialogFromState,
 } from "@/components/chat/artifact-share-controls";
 import { useArtifactShareControls } from "@/components/chat/use-artifact-share-controls";
-import { classifyArtifactType } from "@/components/soul-tools/artifacts-tab-filters";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
@@ -20,73 +16,10 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import type { ChatArtifactRef } from "@/lib/chat-artifacts";
-import { client } from "@/lib/client";
-import { cn } from "@/lib/utils";
-
-/** Extend icon-sm (28px) to a 40px hit target without overlapping neighbors at gap-3. */
-export const iconActionHitArea =
-  "relative after:absolute after:top-1/2 after:left-1/2 after:size-10 after:-translate-x-1/2 after:-translate-y-1/2";
-
-export function toChatArtifactRef(artifact: ArtifactFile): ChatArtifactRef {
-  return {
-    filename: artifact.filename,
-    mimeType: artifact.mimeType,
-    path: artifact.path || artifact.filename,
-    savedAt: artifact.updatedAt,
-    sizeBytes: artifact.sizeBytes,
-  };
-}
-
-const artifactTimestampFormatter = new Intl.DateTimeFormat(undefined, {
-  dateStyle: "medium",
-  timeStyle: "short",
-});
-
-export function formatTimestamp(value: string): string {
-  try {
-    return artifactTimestampFormatter.format(new Date(value));
-  } catch {
-    return value;
-  }
-}
-
-export function getArtifactDownloadUrl(
-  profileId: string,
-  filename: string
-): string {
-  const query = new URLSearchParams({ path: filename });
-  return `${client.baseUrl}/v1/profiles/${encodeURIComponent(profileId)}/artifacts/content?${query.toString()}`;
-}
-
-export function ArtifactIcon({
-  mimeType,
-  filename,
-  className,
-}: {
-  mimeType: string;
-  filename: string;
-  className?: string;
-}) {
-  const kind = classifyArtifactType({
-    filename,
-    mimeType,
-    path: filename,
-    sizeBytes: 0,
-    updatedAt: "",
-  });
-  const iconClass = cn("size-4 text-muted-foreground", className);
-
-  if (kind === "image") {
-    return <Image02Icon aria-hidden className={iconClass} />;
-  }
-
-  if (kind === "video") {
-    return <Film02Icon aria-hidden className={iconClass} />;
-  }
-
-  return <File02Icon aria-hidden className={iconClass} />;
-}
+import {
+  getArtifactDownloadUrl,
+  iconActionHitArea,
+} from "@/pages/files/files-shared";
 
 export function ArtifactRowMenu({
   profileId,
